techpack: audio: q6voice: Force AVCS CVP version 1 on msm8937 SoC
Force the AVCS version to CVP_VERSION_1 for all of the AVCS
services on msm8937 SoC: these ones cannot execute the
Q6 Service Version command on their Q6 DSP because of older
firmware not supporting it.
We know that for the old firmware the version is always 1, hence
forcing it is totally safe.

Change-Id: I5b170c5b7bfbb3073e8946dc8d446c5f48d8106d
Signed-off-by: Aoihara <veronicadisa6@gmail.com>